@@ -13,9 +13,9 @@ class EsiaServiceTest extends TestCase
     private $provider;
     private $service;
 
-    protected function setUp()
+    protected function setUp(): void
     {
-        $this->provider = $this->getMockBuilder(EsiaProvider::class)->disableOriginalConstructor()->setMethods([
+        $this->provider = $this->getMockBuilder(EsiaProvider::class)->disableOriginalConstructor()->onlyMethods([
             'getRandomState',
             'getAuthorizationUrl',
             'getAccessToken',
@@ -39,12 +39,10 @@ class EsiaServiceTest extends TestCase
         $this->assertEquals('url', $this->service->getAuthorizationUrl('state'));
     }
 
-    /**
-     * @expectedException \UnexpectedValueException
-     * @expectedExceptionMessage Generated and passed states must be same: 1 != 2
-     */
     public function testGetResourceOwnerFails()
     {
+        $this->expectExceptionMessage("Generated and passed states must be same: 1 != 2");
+        $this->expectException(\UnexpectedValueException::class);
         $this->service->getResourceOwner('1', '2', '3');
     }
 
